This commit enhances the performance of the MD5 functionality on ARM64 architecture by optimizing the ROUND3 function in the `md5block_arm64.s` assembly file.

- Refactored the `ROUND3` macro to improve the computation order, introducing a new `ROUND3FIRST` macro to handle the initial calculation more efficiently.
- Optimized the XOR operations in the `ROUND3` macro to reduce unnecessary instructions and improve parallelism within the ARM64 architecture.

Performance testing was conducted on an ARM64 Linux machine using Go's benchmark tool. The benchmarks were run 10 times each to ensure statistical significance. The following results were observed:

| Benchmark             | Old Time (sec/op) | New Time (sec/op) | Change |
|-----------------------|-------------------|-------------------|--------|
| Hash8Bytes-8          | 175.0ns ± 2%      | 175.0ns ± 1%      | ~      |
| Hash1K-8              | 2.065µs ± 0%      | 2.060µs ± 0%      | -0.22% |
| Hash8K-8              | 15.31µs ± 0%      | 15.29µs ± 0%      | -0.11% |
| Hash8BytesUnaligned-8 | 174.0ns ± 1%      | 174.0ns ± 1%      | ~      |
| Hash1KUnaligned-8     | 2.067µs ± 0%      | 2.059µs ± 0%      | -0.41% |
| Hash8KUnaligned-8     | 15.44µs ± 0%      | 15.45µs ± 0%      | ~      |

    In terms of throughput:

| Benchmark             | Old Throughput (B/s) | New Throughput (B/s) | Change |
|-----------------------|----------------------|----------------------|--------|
| Hash8Bytes-8          | 43.58MiB/s ± 2%      | 43.69MiB/s ± 0%      | +0.24% |
| Hash1K-8              | 473.1MiB/s ± 0%      | 474.0MiB/s ± 0%      | +0.20% |
| Hash8K-8              | 510.4MiB/s ± 0%      | 511.0MiB/s ± 0%      | +0.11% |
| Hash8BytesUnaligned-8 | 43.80MiB/s ± 0%      | 43.82MiB/s ± 0%      | ~      |
| Hash1KUnaligned-8     | 472.5MiB/s ± 0%      | 474.3MiB/s ± 0%      | +0.38% |
| Hash8KUnaligned-8     | 506.1MiB/s ± 0%      | 505.8MiB/s ± 0%      | ~      |

When testing with large files (e.g., a 3GB file), the runtime was reduced from 8.65 seconds to 7.39 seconds, resulting in an approximate 9% reduction in execution time. This demonstrates a more significant performance gain when handling larger datasets.

Overall, these optimizations provide modest improvements for small input sizes and more noticeable performance benefits when processing larger files, especially in memory-intensive workloads like file hashing.

This PR will be imported into Gerrit with the title and first
comment (this text) used to generate the subject and body of
the Gerrit change.

**Please ensure you adhere to every item in this list.**

More info can be found at https://github.com/golang/go/wiki/CommitMessage

+ The PR title is formatted as follows: `net/http: frob the quux before blarfing`
  + The package name goes before the colon
  + The part after the colon uses the verb tense + phrase that completes the blank in,
    "This change modifies Go to ___________"
  + Lowercase verb after the colon
  + No trailing period
  + Keep the title as short as possible. ideally under 76 characters or shorter
+ No Markdown
+ The first PR comment (this one) is wrapped at 76 characters, unless it's
  really needed (ASCII art, table, or long link)
+ If there is a corresponding issue, add either `Fixes #1234` or `Updates #1234`
  (the latter if this is not a complete fix) to this comment
+ If referring to a repo other than `golang/go` you can use the
  `owner/repo#issue_number` syntax: `Fixes golang/tools#1234`
+ We do not use Signed-off-by lines in Go. Please don't add them.
  Our Gerrit server & GitHub bots enforce CLA compliance instead.
+ Delete these instructions once you have read and applied them